pyproject.toml
[tool.poe.tasks]
# run with eg `uv run poe fmt`
fmt = "ruff format"
lint = "ruff check --fix"
check = "pyright"
test = "pytest"
# run all the above
all = [ {ref="fmt"}, {ref="lint"}, {ref="check"}, {ref="test"} ]

"ci:fmt" = "ruff format --check"  # fail if not formatted
"ci:lint" = "ruff check"

[project]
name = "text2ics"
version = "0.1.0"
description = "Python library to convert text into a ics file with a streamlit app for interaction"
readme = "README.md"
authors = [
    { name = "Jakob Stender Guldberg", email = "jakob1379@gmail.com" }
]

requires-python = ">=3.13.4"
dependencies = [
    "langchain>=0.3.26",
    "llm>=0.26",
    "promptic>=5.5.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.32.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.4",
    "pytest-datadir>=1.7.2",
]
jga = [
    "ipython>=9.4.0",
]
jga-dev = [
    "black>=25.1.0",
    "dap-python>=0.5.0",
    "debugpy>=1.8.14",
    "flake8>=7.1.2",
    "isort>=6.0.1",
    "mypy>=1.16.1",
    "python-lsp-server[all]>=1.12.2",
    "rich>=14.0.0",
    "ruff>=0.12.2",
]

[tool.uv]
# tell uv we're building a package
# (i.e. something we can distribute for others to use)
# uncomment this to use project.scripts
# package = true

# [project.scripts]
# running `postmodern` will run the `postmodern.main` function
# postmodern = "postmodern.hello:main"

[tool.ruff]
target-version = "py313"
line-length = 80  # use whatever number makes you happy

[tool.ruff.lint]
# you can see the looong list of rules here:
# https://docs.astral.sh/ruff/rules/
# here's a couple to start with
select = [
  "A",    # warn about shadowing built-ins
  "E",    # style stuff, whitespaces
  "F",    # important pyflakes lints
  "I",    # import sorting
  "N",    # naming
  "T100", # breakpoints (probably don't want these in prod!)
]
# if you're feeling confident you can do:
# select = ["ALL"]
# and then manually ignore annoying ones:
# ignore = [...]

[tool.ruff.lint.isort]
# so it knows to group first-party stuff last
known-first-party = ["text2ics"]

[tool.pyright]
venvPath = "."          # uv installs the venv in the current dir
venv = ".venv"          #       in a folder called `.venv`
strict = ["**/*.py"]    # use 'strict' checking on all files
pythonVersion = "3.13"  # if library, specify the _lowest_ you support

[tool.pytest.ini_options]
addopts = "--doctest-modules"
